test_kanalservice/requirements.txt
Django==4.0.6
python-dotenv==0.20.0
psycopg2-binary==2.9.3
requests==2.28.1
pandas==1.4.3
celery==5.2.7
django-celery-results==2.4.0
django-celery-beat==2.3.0
gdown==4.5.1
openpyxl==3.0.10
redis==4.3.4
gunicorn==20.1.0
xmltodict==0.13.0
python-telegram-bot==13.13